requirements.txt
alembic==1.18.3
async-generator==1.10
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorlog==6.10.1
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
dataclasses==0.6
fonttools==4.61.1
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
greenlet==3.3.1
h11==0.16.0
httplib2==0.31.2
idna==3.11
importlib_metadata==8.7.1
japanize-matplotlib==1.1.3
joblib==1.5.3
jpholiday==1.0.3
kiwisolver==1.4.9
lightgbm==4.6.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mypy_extensions==1.1.0
numpy==2.4.1
oauth2client==4.1.3
oauthlib==3.3.1
optuna==4.7.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
selenium==4.40.0
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
zipp==3.23.0
